+++ future/src/system.h
@@ -21,7 +21,7 @@ public:
     inline void SetDrawMapPointCallback(std::function<void(const Eigen::VecVector3d&)> callback) {
         backend->SetDrawMapPointCallback(callback);
     }
-    inline void SetDrawPoseCallback(std::function<void(uint, double, const Sophus::SE3d&)> callback) {
+    inline void SetDrawPoseCallback(std::function<void(uint64_t, double, const Sophus::SE3d&)> callback) {
         backend->SetDrawPoseCallback(callback);
     }
 private:
